refactor(ff-sys): split lib.rs into error_codes, constants, and utils modules
lib.rs (257 lines) mixed module declarations, error code constants,
global FFmpeg constants, initialization, and error utilities in one file,
violating the project convention that lib.rs should contain only module
declarations and pub use re-exports.

Extract into three focused modules:

- error_codes.rs — AVERROR_* constants (EAGAIN, EOF, ENOMEM, EINVAL,
  and platform-specific network errno values); 1 unit test
- constants.rs — AV_NOPTS_VALUE, AVFMT_TS_DISCONT, BUFFERSRC_FLAG_KEEP_REF
- utils.rs — ensure_initialized(), av_error_string(), check_av_error! macro;
  2 unit tests

lib.rs now contains only the bindgen include!, module declarations, and
pub use re-exports. All downstream public paths (ff_sys::ensure_initialized,
ff_sys::av_error_string, ff_sys::error_codes::*, ff_sys::AV_NOPTS_VALUE,
ff_sys::check_av_error!) are preserved unchanged.

Resolves #718
