+++ crates/ff-sys/src/lib.rs
@@ -36,7 +36,7 @@ include!(concat!(env!("OUT_DIR"), "/bindings.rs"));
 #[cfg(docsrs)]
 include!("docsrs_stubs.rs");
 
-// Wrapper modules — only compiled when the real bindings are present.
+// ── Wrapper modules (real bindings only) ─────────────────────────────────────
 #[cfg(not(docsrs))]
 pub mod avcodec;
 #[cfg(not(docsrs))]
@@ -46,211 +46,12 @@ pub mod swresample;
 #[cfg(not(docsrs))]
 pub mod swscale;
 
-use std::ffi::CStr;
-use std::sync::Once;
+// ── Sub-modules ───────────────────────────────────────────────────────────────
+mod constants;
+pub mod error_codes;
+mod utils;
 
-/// FFmpeg initialization guard.
-/// Ensures FFmpeg is initialized exactly once.
-static INIT: Once = Once::new();
-
-/// Ensure FFmpeg is initialized.
-///
-/// This function is idempotent and can be called multiple times safely.
-/// It will only perform initialization once.
-pub fn ensure_initialized() {
-    INIT.call_once(|| {
-        // FFmpeg 4.0+ deprecated av_register_all() and it was removed in later versions.
-        // Modern FFmpeg automatically registers codecs/formats at startup.
-        // This function is kept for potential future initialization needs
-        // (e.g., logging configuration, thread settings).
-    });
-}
-
-/// Convert an FFmpeg error code to a human-readable string.
-///
-/// # Arguments
-///
-/// * `errnum` - The FFmpeg error code (negative value)
-///
-/// # Returns
-///
-/// A string describing the error.
-///
-/// # Safety
-///
-/// This function calls FFmpeg's `av_strerror` which is thread-safe.
-pub fn av_error_string(errnum: i32) -> String {
-    const BUF_SIZE: usize = 256;
-    let mut buf = [0i8; BUF_SIZE];
-
-    // SAFETY: av_strerror writes to the buffer and is thread-safe
-    unsafe {
-        av_strerror(errnum, buf.as_mut_ptr(), BUF_SIZE);
-    }
-
-    // SAFETY: av_strerror null-terminates the buffer
-    let c_str = unsafe { CStr::from_ptr(buf.as_ptr()) };
-    c_str.to_string_lossy().into_owned()
-}
-
-/// Macro to check FFmpeg return values and convert to Result.
-///
-/// # Example
-///
-/// ```ignore
-/// let result = check_av_error!(avformat_open_input(...));
-/// ```
-#[macro_export]
-macro_rules! check_av_error {
-    ($expr:expr) => {{
-        let ret = $expr;
-        if ret < 0 {
-            Err($crate::av_error_string(ret))
-        } else {
-            Ok(ret)
-        }
-    }};
-}
-
-/// Common FFmpeg error codes for convenience.
-pub mod error_codes {
-    /// Resource temporarily unavailable (try again).
-    ///
-    /// `AVERROR(EAGAIN)` is platform-specific:
-    /// - Linux/Windows (MinGW): `EAGAIN = 11`, so `AVERROR(EAGAIN) = -11`
-    /// - macOS/BSD: `EAGAIN = 35`, so `AVERROR(EAGAIN) = -35`
-    #[cfg(any(target_os = "macos", target_os = "freebsd", target_os = "openbsd"))]
-    pub const EAGAIN: i32 = -35;
-    #[cfg(not(any(target_os = "macos", target_os = "freebsd", target_os = "openbsd")))]
-    pub const EAGAIN: i32 = -11;
-
-    /// End of file/stream
-    pub const EOF: i32 = -541_478_725; // AVERROR_EOF
-
-    /// Out of memory
-    pub const ENOMEM: i32 = -12;
-
-    /// Invalid data
-    pub const EINVAL: i32 = -22;
-
-    // ── Network errno values (AVERROR = -errno on POSIX) ─────────────────────
-    //
-    // These are used in ff-decode to map FFmpeg network errors to typed variants.
-    // errno numbering differs across platforms:
-    //   - macOS/BSD: uses BSD errno values
-    //   - Windows UCRT: uses its own POSIX-extension errno table (VS2015+)
-    //   - Linux: standard POSIX errno values
-    //
-    // Windows UCRT errno.h (relevant socket codes, added in VS2015/UCRT):
-    //   ECONNREFUSED=107, EHOSTUNREACH=110, ETIMEDOUT=138, ENETUNREACH=118
-
-    /// Connection timed out (`ETIMEDOUT`).
-    #[cfg(any(target_os = "macos", target_os = "freebsd", target_os = "openbsd"))]
-    pub const ETIMEDOUT: i32 = -60;
-    #[cfg(windows)]
-    pub const ETIMEDOUT: i32 = -138; // UCRT: ETIMEDOUT = 138
-    #[cfg(not(any(
-        windows,
-        target_os = "macos",
-        target_os = "freebsd",
-        target_os = "openbsd"
-    )))]
-    pub const ETIMEDOUT: i32 = -110;
-
-    /// Connection refused (`ECONNREFUSED`).
-    #[cfg(any(target_os = "macos", target_os = "freebsd", target_os = "openbsd"))]
-    pub const ECONNREFUSED: i32 = -61;
-    #[cfg(windows)]
-    pub const ECONNREFUSED: i32 = -107; // UCRT: ECONNREFUSED = 107
-    #[cfg(not(any(
-        windows,
-        target_os = "macos",
-        target_os = "freebsd",
-        target_os = "openbsd"
-    )))]
-    pub const ECONNREFUSED: i32 = -111;
-
-    /// No route to host (`EHOSTUNREACH`).
-    #[cfg(any(target_os = "macos", target_os = "freebsd", target_os = "openbsd"))]
-    pub const EHOSTUNREACH: i32 = -65;
-    #[cfg(windows)]
-    pub const EHOSTUNREACH: i32 = -110; // UCRT: EHOSTUNREACH = 110
-    #[cfg(not(any(
-        windows,
-        target_os = "macos",
-        target_os = "freebsd",
-        target_os = "openbsd"
-    )))]
-    pub const EHOSTUNREACH: i32 = -113;
-
-    /// Network unreachable (`ENETUNREACH`).
-    #[cfg(any(target_os = "macos", target_os = "freebsd", target_os = "openbsd"))]
-    pub const ENETUNREACH: i32 = -51;
-    #[cfg(windows)]
-    pub const ENETUNREACH: i32 = -118; // UCRT: ENETUNREACH = 118
-    #[cfg(not(any(
-        windows,
-        target_os = "macos",
-        target_os = "freebsd",
-        target_os = "openbsd"
-    )))]
-    pub const ENETUNREACH: i32 = -101;
-
-    /// I/O error (`EIO`). Same value on all POSIX platforms.
-    pub const EIO: i32 = -5;
-}
-
-/// Invalid PTS value (no presentation timestamp).
-///
-/// This constant indicates that a frame or packet does not have a valid presentation timestamp.
-pub const AV_NOPTS_VALUE: i64 = i64::MIN;
-
-/// `AVFMT_TS_DISCONT` — `AVInputFormat` flag indicating discontinuous timestamps.
-///
-/// Set for live/streaming formats such as HLS (live playlists), RTMP, RTSP,
-/// MPEG-TS (UDP/SRT), and similar sources. Used to detect live streams after
-/// `avformat_find_stream_info` has been called.
-pub const AVFMT_TS_DISCONT: i32 = 0x0200;
-
-/// `AV_BUFFERSRC_FLAG_KEEP_REF` normalized to `i32` for cross-platform use.
-///
-/// bindgen generates `AV_BUFFERSRC_FLAG_KEEP_REF` as `u32` on Linux/macOS
-/// (pkg-config / Homebrew) but as `i32` on Windows (VCPKG). Use this constant
-/// instead of the raw bindgen symbol when calling `av_buffersrc_add_frame_flags`.
-///
-/// The cfg flag `ffmpeg_buffersrc_flag_u32` is emitted by the build script on
-/// platforms where the bindgen type is `u32` (Linux, macOS).
-#[cfg(ffmpeg_buffersrc_flag_u32)]
-pub const BUFFERSRC_FLAG_KEEP_REF: i32 = AV_BUFFERSRC_FLAG_KEEP_REF as i32;
-#[cfg(not(ffmpeg_buffersrc_flag_u32))]
-pub const BUFFERSRC_FLAG_KEEP_REF: i32 = AV_BUFFERSRC_FLAG_KEEP_REF;
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn test_ensure_initialized() {
-        // Should not panic when called multiple times
-        ensure_initialized();
-        ensure_initialized();
-        ensure_initialized();
-    }
-
-    #[test]
-    fn test_av_error_string() {
-        // Test with a known error code
-        let error_str = av_error_string(error_codes::ENOMEM);
-        // The exact message depends on the system, but it should not be empty
-        assert!(!error_str.is_empty());
-    }
-
-    #[test]
-    fn test_error_codes() {
-        // Verify error codes are negative
-        assert!(error_codes::EAGAIN < 0);
-        assert!(error_codes::EOF < 0);
-        assert!(error_codes::ENOMEM < 0);
-        assert!(error_codes::EINVAL < 0);
-    }
-}
+// ── Re-exports ────────────────────────────────────────────────────────────────
+pub use constants::{AV_NOPTS_VALUE, AVFMT_TS_DISCONT, BUFFERSRC_FLAG_KEEP_REF};
+pub use utils::{av_error_string, ensure_initialized};
+// check_av_error! is re-exported at the crate root automatically via #[macro_export]
